python/openrouter_test/requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
json5>=0.9.6
